file rename to fix conflict in python 3.10 tests
