+++ pyrit/datasets/__init__.py
@@ -1,7 +1,7 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT license.
 
-from .fetch_examples import (
+from .fetch_example_datasets import (
     fetch_many_shot_jailbreaking_examples,
     fetch_examples,
 )
