+++ tests/test_fetch_examples.py
@@ -10,7 +10,7 @@ import pytest
 
 
 # Import the functions to be tested
-from pyrit.datasets.fetch_examples import (
+from pyrit.datasets.fetch_example_datasets import (
     _get_cache_file_name,
     _read_cache,
     _write_cache,
@@ -184,7 +184,7 @@ def test_fetch_examples_with_cache():
 
         # Mock the _fetch_from_public_url function to ensure it is not called again
         with patch(
-            "pyrit.datasets.fetch_examples._fetch_from_public_url", MagicMock(return_value=examples)
+            "pyrit.datasets.fetch_example_datasets._fetch_from_public_url", MagicMock(return_value=examples)
         ) as mock_fetch:
             examples_cached_again = fetch_examples(
                 source=SOURCE_URLS["json"], source_type="public_url", cache=True, data_home=str(data_home)
